Three D-Six - Initial Port; Shoving
Ports the framework for the 3D6 system from Daedalus Dock (first seen in #721).

Kapu has had no part in this porting process; but they were the author of the actual system in place here and as such I've elected to put their name as a co-author.

Stat datums are instanced per-mob but don't autodelete YET. That is tomorrow's problem

Co-Authored-By: Kapu1178 <75460809+kapu1178@users.noreply.github.com>