Grip Rolls
Being grabbed is now also on the diceroll system, with the previous probability-based model being excised. Those proficient in grip should also get the HAUL gauntlet trait later-on but that's a "not yet" sorta deal atm

Odds should be roughly similar if both parties have the same stats but can vary drastically. It is very possible to be completely helpless if both parties' stats are set up in just the right way; which is well and away the point. Your stats matter; dammit.