Humans no longer need surnames by default
This commit mostly exists so I can rant before bed. tl;dr; i tried to port citadel sprinting to help sell the combat a bit more. Unfortunately; I'm at port capacity; clearly; because even trying to rewrite it for /tg/code was a disaster. Did make two new SFX for it; though, so.. that's something?

I think we're going without sprinting for a while; unless another codebase on the same license would be more workable for it or I get inspiration to do a massive ground-up piece of work AGAIN. either's feasible ig