Ports Rogue RotationLocking
This is another big combat milestone! FOV no longer looks quite as 'quirked up'

Unfortunately unlike Daedalus; RT's commit history was lost by nature of being, well - leaked code in violation of it's own license lmao - so I can't accredit whoever made it there. Not too bothered by it for that reason; I suppose; but just noting it. This was surprisingly more workable than Citadel sprinting; I'm assuming because it was built off.. is RT based on /vg/ or yogs? One or the other; the ID cards make it obvious which; I forgor.

I had to make some cuts (mostly to eye intent being toggled on BEHALF OF the player; but still)

Sleeping hasn't been ported yet and won't for a lil; if at all. (If it exists I want it to be voluntary and entirely related to levelling up; ala Oblivion. IK that's not "immersive" but it doesn't punish longform RP as much)

Shuffles some keybinds as well so you may have to reset yours