[MDB Ignore] The Tenets Of Life: Skills; Stats; Dicerolls; & Combat/Movement Overhaul (#2)
* Three D-Six - Initial Port; Shoving

Ports the framework for the 3D6 system from Daedalus Dock (first seen in #721).

Kapu has had no part in this porting process; but they were the author of the actual system in place here and as such I've elected to put their name as a co-author.

Stat datums are instanced per-mob but don't autodelete YET. That is tomorrow's problem

Co-Authored-By: Kapu1178 <75460809+kapu1178@users.noreply.github.com>

* Tooltip Half-Fix; Shove Fix Stats Delete Now

* Fix tooltip

* Backend Work

Adds support for defenders using a different skill than the one rolled for their defensive stats

* Grip Rolls

Being grabbed is now also on the diceroll system, with the previous probability-based model being excised. Those proficient in grip should also get the HAUL gauntlet trait later-on but that's a "not yet" sorta deal atm

Odds should be roughly similar if both parties have the same stats but can vary drastically. It is very possible to be completely helpless if both parties' stats are set up in just the right way; which is well and away the point. Your stats matter; dammit.

* Ports cigrolling from Daedalus; too

Ditto with the initial commit; they've no hand in this port process - however lighters are actually feasible to exist in this setting; so I figured.. fuggit

Co-Authored-By: Kapu1178 <75460809+kapu1178@users.noreply.github.com>

* Further Daedalus Usage Ports

A lot more of their use-cases than I thought would be good to have either as reference or mechanically.

Being confused or floored now kneecaps your prowess. Abusing alcohol can make your grip worse. Enabling combat mode offers a boon of extra strength.

Co-Authored-By: Kapu1178 <75460809+kapu1178@users.noreply.github.com>

* Comments no longer relevant

* Disarm Unit Test Potential Fix

* fug

* knockoff unit test

* Update runtimekeep.dmm

* Update runtimekeep.dmm

* Unarmed Punching

Unarmed isn't it's own skill. It shares retail shelfspace with blunt weaponry. I recognize this makes it a little harder to distinguish 'monks' and 'bluntsmen' but i hope the rest of the systems introduce that nuance

* Fixes selfpunch unit test

* Fixes harmpunch unit test

* Update combat.dm

* Basic Mob Punching

Mirrors the changes to human unarmed combat onto unarmed combat with basic mobs. As a human. Basic mobs rolling soon(tm)

* Basic Mobs now also fall under rolling

* Update mob_chains.dm

* Update mob_chains.dm

* Update mob_chains.dm

* Renames, Retools, Reassociates Each Surpemacy Skill

- Swinging / Blunt weaponry have been rolled in with strength to make way for axes. Unfortunately this cucks my thoughts on how this system would work for archers, but. skull emoji

- Unmerges handicraft (daedalus dock original skill) from grip (now grappling) and makes it a celerity skill. Wowee

- All of guard's usecases have been migrated over to mobility/force, as guard is going to mostly be related to parry and the like; with durability taking up the real "armor skills"

* Elusivity

dummy atm; this will be one of a few restraint-slipping skills + have it's own functions.

* Re-balances Force (lmao)

* fug

* Item Attack Rolling

Fixes forcerolls runtiming in the process. Item attacks roll while attacking carbons & simple/basic mobs, just like unarmed attacks.. meaning supremacy is in a functional prototype state! Woo! Hell yeah!!

This took longer than I'd like to admit, of course. I was really spooked by how to handle send_item_attack_message but it MOSTLY is okay. it looks like. at least. I'm going to have to fix more unittests now awaawawwawaawwa

* Sorts a bunch of existing items into melee categories

There will likely be a lot missed. We'll cross that bridge when we come to it

* guh

* This novaflower one may be more hassle than it's worth

I may just remove novaflowers lol

* Enables Native FOV; Gives fireaxes cleaving

Native FOV was a hard requirement to evoke the vibe combat had. IK /tg/ players fuckin' hated it but RT had the last laugh in that regard IG

* Integrates with weapon descriptions

* Update novaflower_burn.dm

* Update novaflower_burn.dm

* Bye felicia

Call me lazy but we won't even fuckin' have botany in the same form

* Update _unit_tests.dm

* Makes movement weightier

* Humans no longer need surnames by default

This commit mostly exists so I can rant before bed. tl;dr; i tried to port citadel sprinting to help sell the combat a bit more. Unfortunately; I'm at port capacity; clearly; because even trying to rewrite it for /tg/code was a disaster. Did make two new SFX for it; though, so.. that's something?

I think we're going without sprinting for a while; unless another codebase on the same license would be more workable for it or I get inspiration to do a massive ground-up piece of work AGAIN. either's feasible ig

* Ports Rogue RotationLocking

This is another big combat milestone! FOV no longer looks quite as 'quirked up'

Unfortunately unlike Daedalus; RT's commit history was lost by nature of being, well - leaked code in violation of it's own license lmao - so I can't accredit whoever made it there. Not too bothered by it for that reason; I suppose; but just noting it. This was surprisingly more workable than Citadel sprinting; I'm assuming because it was built off.. is RT based on /vg/ or yogs? One or the other; the ID cards make it obvious which; I forgor.

I had to make some cuts (mostly to eye intent being toggled on BEHALF OF the player; but still)

Sleeping hasn't been ported yet and won't for a lil; if at all. (If it exists I want it to be voluntary and entirely related to levelling up; ala Oblivion. IK that's not "immersive" but it doesn't punish longform RP as much)

Shuffles some keybinds as well so you may have to reset yours

---------

Co-authored-by: Kapu1178 <75460809+kapu1178@users.noreply.github.com>